Personality Prediction/requirements.txt
pandas
numpy
scikit-learn
joblib
tqdm
matplotlib
streamlit
plotly
pymupdf
python-docx
